third_party/WebKit/LayoutTests/fast/css/variables/cssom-read-expected.txt

This tests Javascript read access to CSS variables in inline styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.style.var.toString() is "[object CSSVariablesMap]"
PASS div.style.var === div.style.var is true
PASS div.style.cssText is "var-a: apple; var-b: banana;"
PASS div.style.var.size is 2
PASS div.style.var.get("a") is "apple"
PASS div.style.var.get("b") is "banana"
PASS div.style.var.get("nonexistant") is ""
PASS successfullyParsed is true

TEST COMPLETE
